web/requirements.txt
Babel==2.3.4
click==6.6
coreapi==2.0.8
coverage==4.2
debtcollector==1.9.0
Django==1.9.10
django-braces==1.10.0
django-cors-headers==1.2.2
django-debug-toolbar==1.6
django-extensions==1.7.4
django-filter==0.15.3
django-rest-swagger==2.1.0
django-shotgun==0.2
djangorestframework==3.5.3
djangorestframework-gis==0.10.1
drf-extensions==0.3.1
drf-hal-json==0.9.1
drf-nested-fields==0.9.3
elasticsearch==2.4.0
elasticsearch-dsl==2.1.0
factory-boy==2.7.0
fake-factory==0.7.2
first==2.0.1
iso8601==0.1.11
itypes==1.1.0
keystoneauth1==2.15.0
Markdown==2.6.7
monotonic==1.2
msgpack-python==0.4.8
netaddr==0.7.18
netifaces==0.10.5
openapi-codec==1.1.7
oslo.config==3.19.0
oslo.i18n==3.10.0
oslo.serialization==2.14.0
oslo.utils==3.18.0
pbr==1.10.0
pip-tools==1.8.0
positional==1.1.1
psycopg2==2.6.2
pyparsing==2.1.10
pyslack-real==0.6.0
python-dateutil==2.5.3
python-keystoneclient==3.7.0
python-swiftclient==3.2.0
pytz==2016.7
requests==2.11.1
rfc3986==0.4.1
simplejson==3.10.0
six==1.10.0
sphinx-me==0.3
sqlparse==0.2.2
stevedore==1.18.0
swift==1.0.2
tqdm==4.9.0
uritemplate==3.0.0
urllib3==1.18.1
wrapt==1.10.8
